+++ Src/renderer/deferred/GeometryPass.h
@@ -20,7 +20,7 @@ public:
 
     void execute(   glm::mat4 projection_matrix, 
                     std::shared_ptr<Camera> main_camera,
-                    GLfloat window_width, GLfloat window_height, 
+                    GLuint window_width, GLuint window_height,
                     GLuint gbuffer_id, GLfloat delta_time, 
                     std::shared_ptr<Scene>);
